Use movable action on Connector example.  Move to common location (will update @layerstack/svelte-actions impl (event callbacks)
